[Pipeline] Improve/fix latency and pipeline verifiers
Adds a bunch more verification to the `pipeline.scheduled` and `pipeline.latency` ops, as well as fixing the register materialization pass.

The register materialization pass has now been modified to  accomodate arbitrarily nested operations within each stage, which is a superset of the support required for `pipeline.latency`, in case operations in the inner body references values defined outside of its stage.